chore: move constant to another class
